TIP-839: Move locale normalizers to Akeneo\Channel
