+++ src/Akeneo/Channel/Bundle/Resources/config/services.yml
@@ -13,6 +13,27 @@ services:
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
 
+    pim_versioning.serializer.normalizer.flat.locale:
+        class: 'Akeneo\Channel\Component\Normalizer\Versioning\LocaleNormalizer'
+        arguments:
+            - '@pim_catalog.normalizer.standard.locale'
+        tags:
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    pim_enrich.normalizer.locale:
+        class: 'Akeneo\Channel\Component\Normalizer\InternalApi\LocaleNormalizer'
+        arguments:
+            - '@pim_user.context.user'
+        tags:
+            - { name: pim_internal_api_serializer.normalizer }
+
+    pim_api.normalizer.locale:
+        class: 'Akeneo\Channel\Component\Normalizer\ExternalApi\LocaleNormalizer'
+        arguments:
+            - '@pim_catalog.normalizer.standard.locale'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
     pim_catalog.repository.locale:
         class: 'Akeneo\Channel\Bundle\Doctrine\Repository\LocaleRepository'
         factory: 'doctrine.orm.entity_manager:getRepository'
