+++ src/Akeneo/Channel/Component/spec/Normalizer/ExternalApi/LocaleNormalizerSpec.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace spec\Akeneo\Tool\Component\Api\Normalizer;
+namespace spec\Akeneo\Channel\Component\Normalizer\ExternalApi;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Channel\Component\Model\LocaleInterface;
-use Akeneo\Tool\Component\Api\Normalizer\LocaleNormalizer;
+use Akeneo\Channel\Component\Normalizer\ExternalApi\LocaleNormalizer;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
