+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/normalizers.yml
@@ -9,7 +9,6 @@ parameters:
     pim_api.normalizer.attribute.class: Akeneo\Tool\Component\Api\Normalizer\AttributeNormalizer
     pim_api.normalizer.attribute_option.class: Akeneo\Tool\Component\Api\Normalizer\AttributeOptionNormalizer
     pim_api.normalizer.attribute_group.class: Akeneo\Tool\Component\Api\Normalizer\AttributeGroupNormalizer
-    pim_api.normalizer.locale.class: Akeneo\Tool\Component\Api\Normalizer\LocaleNormalizer
     pim_api.normalizer.file.class: Akeneo\Tool\Component\Api\Normalizer\FileNormalizer
     pim_api.normalizer.currency.class: Akeneo\Tool\Component\Api\Normalizer\CurrencyNormalizer
     pim_api.normalizer.association_type.class: Akeneo\Tool\Component\Api\Normalizer\AssociationTypeNormalizer
@@ -91,13 +90,6 @@ services:
         class: '%pim_api.normalizer.attribute_group.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.attribute_group'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_api.normalizer.locale:
-        class: '%pim_api.normalizer.locale.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.locale'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
