+++ src/Pim/Bundle/VersioningBundle/Resources/config/serializer/flat.yml
@@ -11,7 +11,6 @@ parameters:
     pim_serializer.normalizer.flat.product.class:           Pim\Bundle\VersioningBundle\Normalizer\Flat\ProductNormalizer
     pim_serializer.normalizer.flat.product_model.class:     Pim\Bundle\VersioningBundle\Normalizer\Flat\ProductModelNormalizer
     pim_serializer.normalizer.flat.value.class:             Pim\Bundle\VersioningBundle\Normalizer\Flat\ValueNormalizer
-    pim_serializer.normalizer.flat.locale.class:            Pim\Bundle\VersioningBundle\Normalizer\Flat\LocaleNormalizer
     pim_serializer.normalizer.flat.collection.class:        Pim\Bundle\VersioningBundle\Normalizer\Flat\CollectionNormalizer
     pim_serializer.normalizer.flat.datetime.class:          Pim\Bundle\VersioningBundle\Normalizer\Flat\DateTimeNormalizer
     pim_serializer.normalizer.flat.metric.class:            Pim\Bundle\VersioningBundle\Normalizer\Flat\MetricNormalizer
@@ -121,12 +120,5 @@ services:
 
     pim_versioning.serializer.normalizer.flat.file:
         class: '%pim_serializer.normalizer.flat.file.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.locale:
-        class: '%pim_serializer.normalizer.flat.locale.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.locale'
         tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
